DH-6 Create bootstrap job for local testing
As we begin trying to wire data into apache trino for testing,
we must first create a test bench with all necessary software
running. I could have done this one-off, and in my terminal -
however the amount of time that it takes to set up a kafka
broker, schema registry, iceberg catalog, backing postgres
database, and data harness is non-trivial. Let's automate this
all behind a script so that future developers have an easy
starting point for development.