DH-9 Add an optional Avro Schema to Table Definition
In order to provide type information to our query engine,
we can store a schema associated with the table. An
alternate approach might be to arbitrarily choose a schema
associated with one of the data sources, though the downside
there is that we can't easily update schemas across all
data sources atomically.

We start with avro now for simplicity, but we'll add the
option to use many types of schema definitions to simplify
the client experience.