DH-9 Add an optional Iceberg Schema to Table Definition
In order to provide type information to our query engine,
we can store a schema associated with the table. An
alternate approach might be to arbitrarily choose a schema
associated with one of the data sources, though the downside
there is that we can't easily update schemas across all
data sources atomically.

We've already done avro, now let's do iceberg.