DH-13 Update table sources atomically
Up to this point, I had overlooked the fact that sources on a
table needed to be modified individually, which eliminates the
atomicity guarantees set by the data harness.

We need to be able to update many of these atomically, which
is also a key reason that we choose to use a PostgresSQL database
as our backing store.

In this endpoint, we now take in many sources, each of which
could even be for different tables (enabling cross table updates)
. We use postgres transactions to ensure they either all succeed
or all fail, and guarantee data consistency of our tables across
updates between their many sources.