DH-17 Add protocol buffers as a schema type
In this change, we allow users to also specify their table schema
in proto buf. This may be easiest if they're already putting
protobuf records in kafka. Down the line, it may eventually
be worth adding thrift.

NOTE: I encounter a very weird bug in my code where some protobuf
schemas do not seem to play well with the data harness. This
must be a result of either using GRPC or using postgres, I really
cannot track it down. I don't think it is postgres from my own
break-pointing, so my guess is it is a bad response object from
GRPC.