DH-19 Add Kafka Connection Information To Source
Initially, I had hoped that users would be responsible for
setting up their own underlying data sources in their querying
engine of choice.

Unfortunately, Spark does not seem to emit a kafka "catalog", so
querying from a kafka broker seems to involve having a user
create a static view via SQL, which says which brokers, topics,
and partitions to read from. Since it also does not cleanly
integrate with the confluent schema registry, I'm forced to
pass in my own schema.

In the Spark plugin, we'll now have to employ individual kafka
source schema strings in order to deserialize messages using
the from_avro() or from_protobuf() functions. For now, I am
not taking in a schema registry URL, so that I can minimize
dependencies. However, that may prove to provide a cleaner
user experience in the future, at which point we can do so.