DH-20 Add YugabyteDB Source
One of the main missions of the data harness is to also allow
users to add relational database tables as a source, in addition
to message brokers and data lakes.

To add a relational database, it is imperative that said database
can support time travel, so that we can retain atomicity in
switching data/schemas across our various sources. This limits
our options to CockroachDB, YugabyteDB, TiDB, and Spanner
(that I know of). The former two are Postgres wire protocol
compliant, making them appealing choices to teams already
using Postgres.

I was able to build out a prototype CockroachDB connector in
Apache Trino. However, attempting to submit changes to open
source connectors in Trino has proven to be non-trivial (I
have gone back and forth for a simple Hudi change for weeks to
no avail). Instead, I'd rather approach this from choosing
technologies that minimize the surface area of code that I need
to touch in existing query engines. While CockroachDB is still
a somewhat viable option, the way that the JDBC connector works
in Spark is more amenable to Yugabyte's system of setting a
"read timestamp" via a SQL session variable.

So far, YugabyteDB's source in the Data Harness should look
like any other. The DataPopulatorIntegrationTest should serve
as an example of how I envision data pipelines use this database
(i.e. insert first, then get the Database timestamp, then
update the source accordingly).