DH-22 Add Spark Catalog Name + Spark Schema to Iceberg Source
When reading an iceberg source, the catalog will need to
be registered externally. From there, we'll point our
query to the appropriate iceberg table.

To do so, an iceberg source that is read by spark must
have its spark catalog name and namespace set appropriately
in the data harness. This is in contrast to the JDBC and
kafka sources which require creating a view for them on
the data harness end.