DH-24 Load data sources atomically
It is imperative that we load our data sources
using a database transaction. Otherwise, we
risk a race condition where two sources were
updated atomically, but we load the state of
one before the update and the state of the
other after the update. That is not acceptable
and ruins our atomicity guarantees. Using a
transaction guarantees that we don't have
write operations interleaved with our database
reads.