DH-26 Add spark connector
This commit adds the initial implementation
for the data harness spark connector. So far,
it supports Kafka with avro encoded data,
YugabyteDB, and Apache Iceberg.

When asking for a data harness table, we
return a DataHarnessTable, which just acts as
a marker object. We intercept this at planning
time, and rewrite it as the union of the
underlying sources. We project each source
back to the expected data output type.