DH-27 Include Spark Integration Test
In our DataPopulatorIntegrationTest, we now
add a remote spark-connect image. We query
it, properly configured with our sources, to
ensure that our spark connector works as
intended.

A couple of quirks here:
1) To ensure that we're using the locally
built version of the spark plugin, I am
manually copying data harness jars into
this spark image.
2) This means that since we are not resolving
them with maven, any direct dependencies of
the jars will need to be added in the
"packages" section.

I was experiencing issues when developing
locally where jars in my maven m2 would not
update in spark because it uses a separate
jar cache called ivy. This way of doing things
should get around the issue, albeit it can
probably be done with some more caching
(right now I just use a temp file system for
ivy, which effectively caches nothing).

Nonetheless, I think that this is a good
starting point and may hopefully serve as an
okay example of how to get an entire data
harness stack up and running. I'll modify
documentation next.