DH-33 Add postgres with temporal tables extension as data source
Make postgres a data source. This will use the temporal tables
extension as described here:
https://github.com/arkhipov/temporal_tables/tree/master

Fortunately, this code should also work with another
postgres table extension which is written completely in
PgSQL and is therefore even more extensible to other
postgres variants.
https://github.com/nearform/temporal_tables

This extension exposes current rows in a table and former
rows in a "history" table with the time period in which they
were valid. By unioning these sources together and only keeping
the rows that were valid for the data-harness provided
timestamp, we've achieved time travel reads/snapshot consistency
with Vanilla PostgresSQL.