DH-37 Add Trino Connector To Repository
Porting initial Trino connector over to the
DataHarness repository.

In this plugin, we just use a union view with SQL
to very simply resolve the combination of many
data sources. This may change in the future if
we find ourselves needing to provide more
complex semantics when merging the views together.
We use existing kafka, protocol buffers, and iceberg
schema resolution logic to convert our table schemas
to Trino data types.

TODO:
- Add support for postgres
- Add plugin to trino docker image and add trino
query to integration test
- Document in a readme