DH-39 Add DataHarness Jar To Local Trino Image
Following the official Trino Guide:
https://trino.io/docs/current/develop/spi-overview.html
https://trino.io/docs/current/installation/plugins.html#plugins-installation

We use the trino-plugin packaging type to create a
zip of our plugin and its dependencies, and then
extract all of them in the proper directory of our
image. Use trino version 476, since this is the
latest one that has published maven images for
external trino plugin dependencies.